# Improve `Repair(1)` (remove unused vertices) performance