--- a/src/transforms/repair.jl
+++ b/src/transforms/repair.jl
@@ -47,17 +47,24 @@ apply(::Repair{0}, mesh::Mesh) = error("not implemented")
 # --------------
 
 function apply(::Repair{1}, mesh::Mesh)
-  count = 0
+  count = Ref(0) # Ref to avoid boxing from map closure (JuliaLang/julia#15276)
+
+  # ordered
   seen = Int[]
   inds = Dict{Int,Int}()
+
+  # reserve worst case memory need: all vertices are actually used
+  sizehint!(seen, nvertices(mesh))
+  sizehint!(inds, nvertices(mesh))
+
   topo = topology(mesh)
   elems = map(elements(topo)) do e
     elem = indices(e)
     for v in elem
-      if v ∉ seen
+      if !haskey(inds, v)
         push!(seen, v)
-        count += 1
-        inds[v] = count
+        count[] += 1
+        inds[v] = count[]
       end
     end
     ntuple(i -> inds[elem[i]], length(elem))