@juliohm 2025-04-23
I am impressed @halleysfifthinc 👏🏽 

Please keep these performance PRs coming!